tool/requirements.txt
requests
PyYAML
termcolor
